Conflicts:
	src/main/java/tn/esprit/spring/entity/Produit.java

+++ src/main/java/tn/esprit/spring/repository/ReclamationRepository.java
@@ -6,6 +6,6 @@ import tn.esprit.spring.entity.Reclamation;
 
 
 
-public interface ReclamationRepository extends CrudRepository<Reclamation,Long>{
+public interface ReclamationRepository extends CrudRepository<Reclamation,Integer>{
 
 }
